Added as many ads as possible so the unblocked hub does not get shutdown
Dear users, due to the amount of growth we have severely surpassed our bandwidth limit. This means we owe more money than we currently have. In order to make enough money we have added ads and a way to support us. If the Unblocked Hub does not make enough money within a set amount of time it will be shut down permanently.